+++ content/definitions/src/main/scala/SimpleWiktionary.scala
@@ -31,3 +31,41 @@ case class SimpleWiktionary(examples: String,
                             verb: String,
                             numerals: String,
                             pronunciation: String)
+
+object SimpleWiktionary {
+  val sectionNames = List(
+    "Examples",
+    "Noun",
+    "Expression",
+    "Homophones",
+    "Prefix",
+    "Synonyms",
+    "Symbol",
+    "Initialism",
+    "Abbreviation",
+    "Suffix",
+    "Contraction",
+    "Notes",
+    "Conjunction",
+    "Antonym",
+    "Phrases",
+    "Antonyms",
+    "Gallery",
+    "References",
+    "Adverb",
+    "Determiner",
+    "Usage",
+    "Synonym",
+    "Acronym",
+    "Interjection",
+    "Determinative",
+    "Preposition",
+    "Adjective",
+    "Etymology",
+    "Pronoun",
+    "Phrase",
+    "Verb",
+    "Numerals",
+    "Pronunciation"
+  )
+}
